@@ -11,34 +11,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~14.2.0",
-    "@angular/common": "~14.2.0",
-    "@angular/compiler": "~14.2.0",
-    "@angular/core": "~14.2.0",
-    "@angular/forms": "~14.2.0",
-    "@angular/platform-browser": "~14.2.0",
-    "@angular/platform-browser-dynamic": "~14.2.0",
-    "@angular/router": "~14.2.0",
-    "@nrwl/angular": "14.8.6",
+    "@angular/animations": "15.0.3",
+    "@angular/cdk": "^15.0.2",
+    "@angular/common": "15.0.3",
+    "@angular/compiler": "15.0.3",
+    "@angular/core": "15.0.3",
+    "@angular/forms": "15.0.3",
+    "@angular/material": "^15.0.2",
+    "@angular/platform-browser": "15.0.3",
+    "@angular/platform-browser-dynamic": "15.0.3",
+    "@angular/router": "15.0.3",
+    "@nrwl/angular": "15.3.3",
     "ng-mocks": "^14.5.0",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~14.2.0",
-    "@angular-eslint/eslint-plugin": "~14.0.4",
-    "@angular-eslint/eslint-plugin-template": "~14.0.4",
-    "@angular-eslint/template-parser": "~14.0.4",
+    "@angular-devkit/build-angular": "15.0.3",
+    "@angular-eslint/eslint-plugin": "15.0.0",
+    "@angular-eslint/eslint-plugin-template": "15.0.0",
+    "@angular-eslint/template-parser": "15.0.0",
     "@angular/cli": "~14.2.0",
-    "@angular/compiler-cli": "~14.2.0",
-    "@angular/language-service": "~14.2.0",
-    "@nrwl/cli": "14.8.6",
-    "@nrwl/cypress": "14.8.6",
-    "@nrwl/eslint-plugin-nx": "14.8.6",
-    "@nrwl/jest": "14.8.6",
-    "@nrwl/linter": "14.8.6",
-    "@nrwl/workspace": "14.8.6",
+    "@angular/compiler-cli": "15.0.3",
+    "@angular/language-service": "15.0.3",
+    "@nrwl/cli": "15.3.3",
+    "@nrwl/cypress": "15.3.3",
+    "@nrwl/eslint-plugin-nx": "15.3.3",
+    "@nrwl/jest": "15.3.3",
+    "@nrwl/linter": "15.3.3",
+    "@nrwl/workspace": "15.3.3",
     "@types/jest": "28.1.1",
     "@types/node": "16.11.7",
     "@typescript-eslint/eslint-plugin": "^5.36.1",
@@ -49,8 +51,8 @@
     "eslint-plugin-cypress": "^2.10.3",
     "jest": "28.1.1",
     "jest-environment-jsdom": "28.1.1",
-    "jest-preset-angular": "~12.2.2",
-    "nx": "14.8.6",
+    "jest-preset-angular": "12.2.3",
+    "nx": "15.3.3",
     "prettier": "^2.6.2",
     "ts-jest": "28.0.5",
     "ts-node": "10.9.1",
