@@ -6,10 +6,11 @@ import { AppComponent } from './app.component';
 import { BookComponent } from './book/book.component';
 import { DiscountPipe } from './discount.pipe';
 import { ChangeColorDirective } from './change-color.directive';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 @NgModule({
   declarations: [AppComponent, BookComponent, DiscountPipe, ChangeColorDirective],
-  imports: [BrowserModule, HttpClientModule],
+  imports: [BrowserModule, HttpClientModule, BrowserAnimationsModule],
   providers: [],
   bootstrap: [AppComponent],
 })
